feat: add category filter functionality to task list

+++ src/app/shared/task-list/task-list.component.ts
@@ -1,14 +1,16 @@
-import { Component, inject } from '@angular/core';
+import { Component, computed, inject, signal } from '@angular/core';
 import { MatButtonModule } from '@angular/material/button';
 import { MatDialog } from '@angular/material/dialog';
 import { MatIconModule } from '@angular/material/icon';
 import { ITask } from '../../interfaces/task.interface';
 import { TaskService } from '../../services/task.service';
 import { EditTaskDialogComponent } from '../edit-task-dialog/edit-task-dialog.component';
+import { FilterComponent } from '../filter/filter.component';
+import { RemoveTaskDialogComponent } from '../remove-task-dialog/remove-task-dialog.component';
 
 @Component({
   selector: 'app-task-list',
-  imports: [MatButtonModule, MatIconModule],
+  imports: [MatButtonModule, MatIconModule, FilterComponent],
   templateUrl: './task-list.component.html',
   styleUrl: './task-list.component.scss',
 })
@@ -17,6 +19,15 @@ export class TaskListComponent {
   dialog = inject(MatDialog);
 
   tasks = this.taskService.tasks;
+  selectedCategory = signal<string>('all');
+  filteredTasks = computed(() => {
+    if (this.selectedCategory() === 'all') {
+      return this.tasks();
+    }
+    return this.tasks().filter(
+      (task) => task.category === this.selectedCategory()
+    );
+  });
 
   toggleCompletion(title: string, category: string): void {
     this.taskService.toggleTaskCompletion(title, category);
@@ -36,6 +47,18 @@ export class TaskListComponent {
   }
 
   onRemoveTask(title: string, category: string): void {
-    this.taskService.removeTask(title, category);
+    const dialogRef = this.dialog.open(RemoveTaskDialogComponent, {
+      width: '400px',
+    });
+
+    dialogRef.afterClosed().subscribe((result: boolean) => {
+      if (result) {
+        this.taskService.removeTask(title, category);
+      }
+    });
+  }
+
+  onFilterChange(category: string): void {
+    this.selectedCategory.set(category);
   }
 }
